Added section with hyperv-specific options. The differencing disk should significantly speed up this step. Increased the IP timeout because on some test boxes this took longer than 2 minutes.
Updated with check for Hyper-V with logic and variables to target this if present. It will still fall back to virtualbox if that is found and Hyper-V is not.

Added section to install Hyper-v tools from the vmguest.iso on first boot. This command will just error out for non-Hyper-V installs where it isn't present.

moved vm guest tool installation to the initial floppy

Added hyper-v first boot section in vm-guest-tools, add this script to the floppy drive during the hyper-v build, and call it with autounattend

Resolving pull request issues