Range of IP address of the linux machine invalid.
The range of the IP address of the metasploitable Linux virtual machine should be between 192.168.56.0 and 192.168.56.21.  Otherwise a error occurs as following...

```
The IP address configured for the host-only network is not within the
allowed ranges. Please update the address used to be within the allowed
ranges and run the command again.

  Address: 172.28.128.3
  Ranges: 192.168.56.0/21

Valid ranges can be modified in the /etc/vbox/networks.conf file. For
more information including valid format see:

  https://www.virtualbox.org/manual/ch06.html#network_hostonly

```